move Player and Scraper to their own files and make a few improvements to error checking and variable placement
